requirements.txt
Django==3.1.7
jupyter==1.0.0
django-extensions==3.1.1
django-ckeditor==6.0.0
requests==2.25.1
django-heroku==0.3.1
gunicorn==20.0.4
whitenoise==5.2.0
djangorestframework~=3.12.2
django-filter==2.4.0
drf_extensions==0.7.0
django-cors-headers==3.7.0
djangorestframework-simplejwt==4.6.0